Fix repo links in pom.xml and restructure packages
